+++ src/meme_4.6.0/src/dir.h
@@ -1,3 +1,0 @@
-#define MEME_DIR "/net/noble/vol1/home/cegrant/meme"
-#define ETC_DIR "/net/noble/vol1/home/cegrant/meme/etc" 
-#define BIN_DIR "/net/noble/vol1/home/cegrant/meme/bin"