+++ src/meme_4.6.0/src/glam2_init.h
@@ -1,8 +1,0 @@
-#ifndef GLAM2_INIT_H
-#define GLAM2_INIT_H
-
-#include "glam2_glam2.h"
-
-void init(data *d);
-
-#endif