+++ src/meme_4.6.0/website/html/showhide.js
@@ -1,9 +1,0 @@
-function showhide(blockid,textid) {
-    if (document.getElementById(blockid).style.display == "none") {
-        document.getElementById(blockid).style.display = "block";
-        document.getElementById(textid).innerHTML = "-";
-    } else {
-        document.getElementById(blockid).style.display = "none";
-	document.getElementById(textid).innerHTML = "+";
-    }
-}