+++ test/test_tools.py
@@ -18,12 +18,14 @@ class TestMotifProgram(unittest.TestCase):
         return os.path.exists(bin) and os.access(bin, os.X_OK)
     
     def ap1_included(self, motifs):
+        #if len(motifs) == 0:
+        #    return False
         ap1 = motif_from_consensus("TGASTCA")
         mc = MotifComparer()
-        for motif in motifs:
-            match = mc.get_closest_match(ap1, motif)
-            if match["TGASTCA"][1][3] < 1e-6:
-                return True
+        match = mc.get_closest_match(ap1, motifs, metric="seqcor")
+        print(match)
+        if match["TGASTCA"][1][0] >= 0.8:
+            return True
         return False
 
     def test_tools(self):
@@ -41,6 +43,7 @@ class TestMotifProgram(unittest.TestCase):
                     "xxmotif", # takes too long
                     "trawler", # unpredictable, sometimes doesn't find the motif
                     "weeder", # doesn't work at the moment
+                    "posmo", # motif doesn't predictably look like AP1
                     ]:
                 continue
            
@@ -51,9 +54,6 @@ class TestMotifProgram(unittest.TestCase):
                     "hms",
                     "improbizer",
                     "motifsampler",
-                    "posmo",
-                    "meme", # until the bioconda package is fixed
-                    "memew", # until the bioconda package is fixed
                     ]:
                     continue
 
@@ -61,6 +61,9 @@ class TestMotifProgram(unittest.TestCase):
             print("Testing {}...".format(t))
             
             (motifs, stderr, stdout) =  t.run(self.fa, params)
+            print(motifs)
+            print(stderr)
+            print(stdout)
             self.assertTrue(self.ap1_included(motifs))
 
     def tearDown(self):